ocr-service/requirements.txt
modal>=0.64.0
pillow>=10.1.0
